Added setup commands and option to specify device
The script previously did not configure any framework settings for this
script, so I added commands to disable any features that could affect
the performance numbers like wifi, NFC, screen rotation. It also clears
notifications and starts the device on the home screen.

It's not possible prior to this commit to run the app-launcher on a
host with multiple devices attached to adb, so I added a '-s <serial
number' option that allows the user to specify a specific device based
on its serial number.

I also changed the spacing of the script to be all spaces instead of
mixed tabs and spaces.

Test: Manually ran script on a marlin device

Change-Id: I82c621fe280033c55ba38bfdf95836fd95f5ed58