Add an option to explicitly specify blockdev.
Add a -b option, that allows user to explicitly specify a blockdev
that we want to get stats out of. For example, this allows us to
specify the userdata partition (on which IOshark runs). Or in the rare
case where support for a new device has not been added yet, we can
still run IOshark on it by using this option.

Test: Run IOshark with the -b option passed in and verify that the
device stats reported match with /proc/diskstats.

Change-Id: Ic835697a6aae4484c95537e7d489a9c1fefc0110
Signed-off-by: Mohan Srinivasan <srmohan@google.com>