IOshark portability across 32- and 64-bit, little/big endian devices.
Make IOshark portable across 32-bit and 64-bit devices, as well as
little endian and big endian architectures. This is achieved by
locking down the on-disk format of IOshark workload files. We can now
generate the workload files on a 64-bit host and use those workload
files to run the benchmark on a 32-bit userspace device (such as
fugu). The change also has minor fixes to make IOshark work on taimen
and walleye.

Test: Test across 64- and 32-bit devices. Generate workload files on a
64 bit device and test on 32-bit devices (as well as 64-bit devices).

Change-Id: I700ca5ae0da0ab029d977c906a63f1cd20a78a14
Signed-off-by: Mohan Srinivasan <srmohan@google.com>