simpleperf: adjust sample freq instead of checking it.
Linux kernel introduces methods to adjust max sample rate
based on load with patch "perf: Drop sample rate when sampling is too slow".
It is annoying if the first time recording with -f xxx option successes,
but following runs fail because the max sample rate is decreased.
So instead of checking if sample_freq <= max_sample_rate, I prefer to
change it to adjust sample_freq in range of [1, max_sample_rate].

Test: run simpleperf_unit_test.
Change-Id: Ic2e93b7fb861c704a4aa8199d08b5b0650db2c23