build_verity_metadata.py: allow disabling verity
Allowing building metadata with the "VOFF" magic, for ENG builds to
disable dm-verity by default. The resulting metadata is the same as
triggering an "adb disable-verity" on an USERDEBUG image.

With this, fs_mgr doesn't need to know whether it's an ENG build and
can rely on the magic to know it should enable dm-verity or not.

Bug: 63056044
Test: boot sailfish eng/userdebug builds
Change-Id: Ie81de372aa057f3aafa0eb571b460e73aa5eb375