Make the NOT_CONFIG_SYSVIPC test also exercise the syscalls
This change augments the NOT_CONFIG_SYSVIPC test, such that in addition
to being evidence-based (by inspecting some paths in /proc), it also
tries to invoke the syscalls.  This is done because in some platforms
like Chrome OS, the SYSVIPC kernel config is enabled (because some parts
of the system require that to boot), but when Android is running, all
SYSVIPC syscalls are blocked by an LSM in the kernel.

It also changes the /proc/sysvipc path check from using access(2) and
F_OK to R_OK, since the paths itself is present and visible.

Bug: 77490033
Test: CtsKernelConfigTestCases
Change-Id: I79df1816e0e5d3618da2a0e242c3d685352b7220