Convert mkuserimg_mke2fs.sh to python
The mkuserimg_mke2fs.sh was a shell script, which expects all the
arguments in order. This cl changes the script to python which leverages
python's argumentparser and unittest modules.

The script usage is unchanged; and this tool will be packed into the
otatools.zip in the follow up cls. Then we'll change the caller site
gradually; and remove the old shell script after that.

Bug: 112555072
Bug: 63866463
Test: run unit tests & build a userdata image
Change-Id: Ie6b687da3de31a3481363f01d2b5c12df91ca5ce