Remove memtest.
This is really only useful for 32 bit devices, and it's not all that useful.
It was designed to check the different memory instructions and indicate which
would likely give the best results for memcpy/memset.

However, we aren't doing any active performance improvements for these 32 bit
routines any more and it can be confusing that this is 32 bit only, so it's
time to go.

Test: NA
Change-Id: Ib56a2ca0a603ac9477422431c3c13f3112366473