Do not create property_watch_loop thread for zygote
Bug: http://b/116873221
Bug: http://b/133872559

Do not create thread running property_watch_loop for zygote (which is
essentially /system/bin/app_process invoked with a specific argument to
start the zygote).

The reason this is needed is because when the zygote forks system_server
or an app, it waits for all threads to stop.  But the thread created
here doesn't know that it has to stop.  So zygote gets stuck waiting and
the device doesn't boot.

This check is only needed for the platform, but can be done on any
version after Android L, when getprogname() was added.

Test: cuttlefish with coverage enabled can boot.
Change-Id: I65aa603a88bf8da1f14b5c4ada3adf3776f33275
(cherry picked from commit cbaf58206d8500461a222a2cb94a6b8b421ad419)