mke2fs.conf: Don't include resize_inode by default.
This flag cause mke2fs to reserve a percentage of blocks as metadata for
resizing the filesystem. This wastes a small amount of space (a few
hundred KB), but more importantly, makes incremental diffs worse when
using BaseFS. Since the number of reserved blocks can change with small
differences to the image size, this can cause off-by-one issues leading
to large sequences of moves.

In a sample Virtual A/B simulation, this reduces the snapshot size by
200MB.

Bug: 139201772
Test: debugfs -R stats <image>
Change-Id: I57c1456356975dddbb5bcc12ea25ded189848568