Add --comm filter to report_sample.py
This is modelled off the --comm flag in pprof_proto_generator.py.

It's also similar to `perf script --comms`. Perhaps I should add an alias

I had to also move the filename from being a positional argument to
a -i/--record_file argument, because otherwise if I ran

$ report_sample.py --comm DefaultPool1 my_perf.data

Then my_perf.data was interpreted as an additional comm argument.

The -i brings this tool into line with the other tools too.
No docs needed to be updated -- they all use the implicit perf.data default.

Also add a few tests.

This is particularly useful for filtering to the UI thread for
visualising in Flamescope.

BUG=200118152

Change-Id: Idd6dc5a85fa60156bab30aca672c3cd69c402119
Test: Ran test/test.py -p TestReportSample.\*