simpleperf: fix missing python test results.
When some tests exceed MAX_TRY_TIME, their results are not written
to TestSummary. So fix it in wait_for_test_results(). Also keep
all tests names in TestSummary. So we can show test results in
TestSummary even if they are not received.

Also improve progress bar for restarted TestProcess.

Bug: 203246914
Test: run scripts/test/test.py

Change-Id: I6d903f10c14321e40e3b4ddd700b8953a7eb4721