simpleperf: Store perf_event_attr value in EventAttrWithId
Currently we use "const perf_event_attr*" field in EventAttrWithId.
Because no users want to modify the field, and copying the pointer
is more efficient than copying the value, which takes 128 bytes.

But now there is a case that we want to modify perf_event_attr.
It's more straightforward to store the value instead of the pointer.

The change also makes it more convenient to store EventAttrIds directly
in RecordFileReader.

Bug: 273341791
Test: run simpleperf_unit_test
Change-Id: I47026cace790c45d18bfca16d1955b80aac41b87