Implemented argument validation for app-startup and multiple runs.
Before implementing the app-startup event, we need to make sure
--app is even passed. To run multiple tests, we also need to
make sure the --ui is disabled. 2 checks are added in the
verify_args method of torq.py to handle these 2 edge cases.
Furthermore, 3 new unit tests and a constant are added to
tests/unit_test_torq.py to test this new logic, and minor
changes are made to a few existing tests, including name
and argument changes.

Test: atest torq_unit_test
Bug: 362550220
Bug: 360924873

Change-Id: I269e9fca77dde439e03d95c4bb3993e8eccd5df1