Add a present bytes parameter in traces.
In order to more accurately rerun traces and contain more information
about how allocations are used, keep track of the total present bytes
for the allocations. These present bytes represent the amount of the
allocation that has been written by the program that is being traced.

Update the verify trace code to make sure that the present bytes is
never larger than the size of the allocation.

Test: All unit tests pass.
Change-Id: I8e4dc0c9774dff696cc0448889a3d2f337606647