Fix static linkage in libcutils_trace_bindgen
A bug in Soong was causing static libraries to propagate when they
shouldn't have. As a result, this was propagating the
libcutils_trace_bindgen static lib when it was not being included as a
whole_static_lib.

Switch libcutils_trace_bindgen from static_libs to whole_static_libs so
that it gets bundled with libcutils_trace_bindgen as intended.

Bug: 377932175
Test: m libatrace_rust
Change-Id: I571749bb021966985915a86a8e5a10fb47e90f7b