Add a verify_trace tool.
Create a tool that can verify that a trace is valid.

In addition, add a method to attempt to repair trace files. There is a
small chance that a race can occur where two allocations entries are
written to two different threads that return the same pointer. This can
happen if the free is occuring on a different thread and gets written
out after both other allocations.

Test: Ran this on traces that are bad and good and verified the results.
Test: Verified a repair works.
Change-Id: Ibcf58c598935ebf4410eec55c18ad47420c6f126