Fix bugs in torq
User switch currently waits for the duration the user set after doing a
user switch before starting a trace. This is fine for traces of a few
seconds, but is a problem if the user sets a duration of a few minutes
or hours, because they will have to wait that amount of time before the
trace starts. Change this to wait for a fixed amount of time.
Additionally, there is a race condition in BootCommandExecutor between
device.reboot and device.wait_for_device. When reboot returns, the
device is still available for a couple of seconds, so
device.wait_for_device will sometimes return immediately before reboot
actually starts. Change so reboot only returns when the device serial is
no longer in the list of device serials.

Test: m + presubmit
Fixes: 380324635
Fixes: 380324944
Change-Id: I9000a0f6f8e0557cd08084ba7c79c3a5babcd132