pprof: apply to outermost frame when reusing function name from report library
The logic in question was added to preserve ProGuard mappings in
combination with symbols from the ODEX file in binary cache. This
however caused issues with inlined frames (from non-ODEX native
libraries) because the function name was copied against the first frame,
which was the innermost (inlinee) frame. The correct target is the
outmost (inliner) frame, so this led to duplicated and corrupt stack
traces. Fix this.

Bug: 376631018
Test: compared pprof results before/after (see Bug)
Change-Id: I52e09cdcd8796789fc21e5291ed46bd6da3c78b3