simpleperf : enhanced profiling capabilities utilizing the Decode Stream Buffer (DSB) event types for Intel x86 Core and Atom processors
As mentioned in https://perfmon-events.intel.com/ updated the core and atom DSB events with Event Sel and UMask Bits. In order to analyze the DSB coverage and optimize the utilization for x86 platforms, these  events will help in getting the insights for the performance bottlenecks and help in understanding how   efficiently we can optimize the DSB coverage.

Bug: 400832666
Test: Run simpleperf
Change-Id: Idf0f914607de65f5db187c754ff64207450c2665
Signed-off-by: Priyanka Bose <priyanka.bose@intel.com>